+++ src/main/kotlin/com/nylas/models/FindContactQueryParams.kt
@@ -1,0 +1,8 @@
+package com.nylas.models
+
+import com.squareup.moshi.Json
+
+data class FindContactQueryParams(
+  @Json(name = "profile_picture")
+  val profilePicture: Boolean? = null,
+): IQueryParams