+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 * Added support for sending drafts
+* Added support for the contacts API
 
 ### Changed
 * Fixed issue with sending scheduled messages